pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "unifiedcatalogpy"
dynamic = ["version"]
dependencies = [
  "requests>=2.28.0",
  "azure-identity>=1.12.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "black>=22.0.0",
  "isort>=5.10.0",
  "flake8>=5.0.0",
  "mypy>=1.0.0",
]
test = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
]
description = "A Python client for interacting with the Microsoft Purview Unified Catalog API."
authors = [
    { name = "Olaf Wrieden" }
]
license = { file = "LICENSE" }
keywords = ["Microsoft Purview", "Unified Catalog", "Purview", "Data Governance", "Data Catalog"]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/olafwrieden/unifiedcatalogpy"
Documentation = "https://github.com/olafwrieden/unifiedcatalogpy"
Repository = "https://github.com/olafwrieden/unifiedcatalogpy.git"
"Bug Tracker" = "https://github.com/olafwrieden/unifiedcatalogpy/issues"
Changelog = "https://github.com/olafwrieden/unifiedcatalogpy/commits/main/"

[tool.setuptools]
packages = ["unifiedcatalogpy"]

[tool.setuptools.dynamic]
version = { attr = "unifiedcatalogpy.__version__" }